Add possibility to define custom code for cooling
These changes allow to define custom code to be run when using the _cooldown_ preset. It makes the _cooldown_ preset similar to the _other_ ones. Note that this is something that can be done on Mainsail.

Here an example in the configuration:
```
[preheat cooldown]
gcode: 
    TURN_OFF_HEATERS
    M107
```

Note that the cooldown could be made a preset like any other? I'd be happy to take a look at it if wanted!