Update zcalibrate.py
Proposal: If abort is pressed, force auto home to reposition printhead when Z calibrate is restarted. 

Issue: In the case that the user aborts, the machine does not rehome. So, worst case, the printhead is off the bed on custom applications or in a place that the bltouch cannot probe, so when you press Start, it tries to home z in that position and moves below the gantry limits (corexy). Something will bend or break.