Xserver-XSDL cmd args to stop screen blanking (#794)
Added documentation to stop Xserver-XSDL from enabling its 5 minute default screen saver and blanking the screen.  In this scenario even after enabling the 'stay away' option in the debugging or developer options of the android device, the screen will still go black, but leave the back light on.  To stop this this, you must disable the Xserver screen saver option in the start up arguments of the server app itself.  I spend more than a few hours looking for and figuring this out, when this little bit of info would have saved me a lot of time.

Not sure if where I placed it is the best place in the doc, but the info is sound.  I am using my android over USB, FYI.