refactor(job_status): time estimation
Even if we are very confident in the estimate, such as when there is previous historic data,
variables like speed factor and possible excluded objects can impact the accuracy.
Taking these cases into account seems to be tricky.
To address this, I revised the  weightings of my previous formula based on practical considerations and observations.