feat: Allow buttons to be focused
This allows elements to be focused by pressing tab or shift-tab, and
selected with enter or space.

The motivation for this is to be able to use a rotary encoder on my LCD
display panel to navigate KlipperScreen. A bit more background and an
example video is at
https://github.com/KlipperScreen/KlipperScreen/discussions/1549

I haven't found any side effects which otherwise impacts touchscreen
use.